feat: :sparkles: sync basic development files to template (#17)
# Description

Updated the `justfile` so that we can copy the basic dev files
(`.gitignore`, etc) in the parent folder to the template. That way we
only need to update one of them if we make changes to them. And these
files shouldn't like be different between the parent and the template
folder.

This PR needs a quick review.

## Checklist

- [x] Ran `just run-all`