test: :white_check_mark: improve robustness of the tests (#49)
# Description

Created a bash function to minimize duplicate code, put the individual
tests in `( )` to contain their effects, always `cd` into the test dir,
and run `just run-all` at the end of each test within the test dir to
check that the recipes work too.

Closes #21

This PR needs an in-depth review.

## Checklist

- [x] Ran `just run-all`