refactor: :recycle: revise so `github_repo` is name of folder (#60)
# Description

I removed the use of `website_abbrev` (it was only used once anyway).
Also found a bug. We need to use `realpath` for `dst_path` otherwise it
will use `.` if the template destination path is the same as the working
directory. I also split it so the `github_repo` was determined from
`dst_path` (which requires use of the `realpath` first).

Closes #20

This PR needs an in-depth review.

## Checklist

- [x] Ran `just run-all`

---------

Co-authored-by: martonvago <57952344+martonvago@users.noreply.github.com>