fix: :bug: `github_repo` needs to be a question to store properly (#92)
# Description

I found a bug with how copier does updating. Seems we can't use
`dst_path` for `when: false` stored values, as it reverts to using the
temporary location when running `copier update`. So this moves it up as
an actual question to be stored in the answer file.

This PR needs a quick review.

## Checklist

- [x] Ran `just run-all`