refactor: :recycle: generate `_contributors.yml` into the `includes/` folder (#90)
# Description

Revise the get contributors file to output text and redirect to the
repos `includes/` folder. Originally it wasn't meant to be kept in the
repo (ignored by Git), but since the build is a bit weird with it
included, this resolves that by keeping it as a committed file.

This PR needs a quick review.

## Checklist

- [x] Formatted Markdown
- [x] Ran `just run-all`

---------

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>